requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
websockets==12.0
pydub==0.25.1
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
aiofiles==23.2.1 